refactor(chat): extract chat_manager/deps_permission.rs (#292)
Task 10 of the split plan: PlanHostImpl and PhDeps (and their
PlanHost/PermissionHandlerDeps impls) move into their own file,
owning their own construction via a build() helper the constructor
now calls.
